requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.15
sqlalchemy[asyncio]>=2.0.15
pydantic>=2.3.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.12.0
asyncpg>=0.28.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.24.0
requests>=2.31.0
